schedule: cache .after()
the reason why this was rather slow with many occurrences in one month
(in my test ~60500) is that to find recurrences they *have to be built
iteratively*. You *cannot* "jump into" the stream of occurrences
at an arbitrary point in time, you'll have to compute every occurrence
before that. so either "if .between()" or ".after()" have to compute all
occurrences up to the end point / first occurrence.

so caching these intermediary results improves perf by 30-40x, as one
would expect (for week in month: for day in week:).